rag_system/requirements.txt
# Core RAG Dependencies
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
groq==0.11.0

# Vector Store & Embeddings
chromadb==0.5.11
sentence-transformers==3.1.1

# API Framework
fastapi==0.115.0
langserve==0.3.0
uvicorn[standard]==0.30.6

# Data Processing
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
redis==5.0.8

# Development & Monitoring
langsmith==0.1.129
pytest==8.3.3
pytest-asyncio==0.24.0

# Prometheus Monitoring
prometheus-client==0.21.0